Squashed commit
Add test files with long int IDs

The test-large-IDs files use the new
"long int" IDs which break the current code.

add failing test

Remove track_oracle assumption that IDs are unsigned ints.

This commit converts all explicit use of unsigned as the data type
for external_id to the data_term's tracking::external_id::Type.

Convert KPF and track_oracle ID type to "unsigned long long".

This commit switches the type for IDs from unsigned to unsigned long long
in both KPF and track_oracle.

Fix metadata parsing.

This fix generalizes metadata parsing to handle both the "official"
payload of a string:

   meta: "this is a metadata string"

...and the D&A variant of a YAML structure:

   meta: { team: kitware }

Add a test for reading KPF activities.

This test first reads the geometry file, then the activity file. The
process of reading the activity file verifies that all actors mentioned
in the activity file can be located in the geometry file.

+++ arrows/kpf/yaml/kpf_canonical_types.h
@@ -76,8 +76,8 @@ struct KPF_YAML_EXPORT bbox_t
 struct KPF_YAML_EXPORT id_t
 {
   enum {DETECTION_ID=0, TRACK_ID, EVENT_ID };
-  size_t d;
-  explicit id_t( size_t i ): d(i) {}
+  unsigned long long d;
+  explicit id_t( unsigned long long i ): d(i) {}
   id_t(): d(0) {}
 };
