+++ track_oracle/data_terms/data_terms.h
@@ -76,7 +76,7 @@ namespace detection {
 
 namespace tracking {
 
-  DECL_DT( external_id, unsigned, "track ID; unique within a session but not a UUID" );
+  DECL_DT( external_id, unsigned long long, "track ID; unique within a session but not a UUID" );
   DECL_DT( timestamp_usecs, unsigned long long, "timestamp of a frame, in usecs; epoch is data-dependent" );
   DECL_DT( frame_number, unsigned, "frame number; relationship to any downsampling is unspecified" );
   DECL_DT( fg_mask_area, double, "area of foreground mask; in pixels?" );
@@ -101,7 +101,7 @@ namespace tracking {
 } // ...tracking
 
 namespace events {
-  DECL_DT( event_id, unsigned, "event ID; unique within a session but not a UUID" );
+  DECL_DT( event_id, unsigned long long, "event ID; unique within a session but not a UUID" );
   DECL_DT_RW_STRXMLCSV( event_type, int, "event type: currently always in the VIRAT domain" );
   DECL_DT( event_probability, double, "event probability" );
   DECL_DT_RW_STR( source_track_ids, std::vector<unsigned>, "Track IDs contributing to the event" );
